Clean up tests around warnings
